threaded_search now testing direct links and search queries
we're returning an object that has both search queries that google
recommends and direct links for search results. This object (hash) has
both and not we're prepped for using these recommended searches from
google (hopefully other search engines work the same way) in a threaded,
parallel way async to follow the links and see what we get from those.